Conflicts:
	tests/app/dao/test_notification_dao.py

+++ tests/conftest.py
@@ -52,7 +52,8 @@ def notify_db_session(request):
     def teardown():
         db.session.remove()
         for tbl in reversed(meta.sorted_tables):
-            db.engine.execute(tbl.delete())
+            if tbl.name not in ["provider_details"]:
+                db.engine.execute(tbl.delete())
         db.session.commit()
 
     meta = MetaData(bind=db.engine, reflect=True)
